+++ src/main/java/ccommit/stylehub/product/controller/ProductController.java
@@ -7,6 +7,8 @@ import ccommit.stylehub.product.dto.request.StockUpdateRequest;
 import ccommit.stylehub.product.dto.response.ProductCursorResponse;
 import ccommit.stylehub.product.dto.response.ProductOptionResponse;
 import ccommit.stylehub.product.dto.response.ProductResponse;
+import ccommit.stylehub.product.enums.MainCategory;
+import ccommit.stylehub.product.enums.SubCategory;
 import ccommit.stylehub.product.service.ProductService;
 import ccommit.stylehub.user.enums.UserRole;
 import jakarta.servlet.http.HttpServletRequest;
@@ -19,27 +21,45 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
  * @author WonJin Bae
  * @created 2026/03/25
+ * @modified 2026/03/27 by WonJin - feat: 커서 기반 전체 상품 목록 조회 API 추가
+ * @modified 2026/04/01 by WonJin - refactor: ProductViewController를 ProductController로 통합
  *
  * <p>
- * STORE 역할 사용자의 상품 등록, 내 스토어 상품 조회, 재고 관리 API를 제공한다.
+ * 상품 관련 API를 제공한다.
+ * 공개 API(목록/단건 조회)는 인증 불필요, 스토어 API(등록/재고 관리)는 STORE 역할이 필요하다.
  * </p>
  */
 @RestController
-@RequestMapping("/api/v1/stores/{storeId}/products")
 @RequiredArgsConstructor
-@RequiredRole(UserRole.STORE)
 public class ProductController {
 
     private final ProductService productService;
 
-    @GetMapping
+    //공개 API (비인증)
+    @GetMapping("/api/v1/products")
+    public ResponseEntity<ProductCursorResponse> getProducts(
+            @RequestParam(required = false) Long cursor,
+            @RequestParam(required = false) Long storeId,
+            @RequestParam(required = false) MainCategory mainCategory,
+            @RequestParam(required = false) SubCategory subCategory,
+            @RequestParam(required = false) Integer size) {
+        return ResponseEntity.ok(productService.getProducts(cursor, storeId, mainCategory, subCategory, size));
+    }
+
+    @GetMapping("/api/v1/products/{productId}")
+    public ResponseEntity<ProductResponse> getProduct(@PathVariable Long productId) {
+        return ResponseEntity.ok(productService.getProduct(productId));
+    }
+
+    //  스토어 API (STORE 권한 필요)
+    @GetMapping("/api/v1/stores/{storeId}/products")
+    @RequiredRole(UserRole.STORE)
     public ResponseEntity<ProductCursorResponse> getMyStoreProducts(
             @PathVariable Long storeId,
             @RequestParam(required = false) Long cursor,
@@ -49,7 +69,8 @@ public class ProductController {
         return ResponseEntity.ok(productService.getMyStoreProducts(userId, storeId, cursor, size));
     }
 
-    @PostMapping
+    @PostMapping("/api/v1/stores/{storeId}/products")
+    @RequiredRole(UserRole.STORE)
     public ResponseEntity<ProductResponse> registerProduct(
             @PathVariable Long storeId,
             @Valid @RequestBody ProductCreateRequest request,
@@ -59,7 +80,8 @@ public class ProductController {
                 .body(productService.registerProduct(userId, storeId, request));
     }
 
-    @PatchMapping("/{productId}/options/{optionId}/stock")
+    @PatchMapping("/api/v1/stores/{storeId}/products/{productId}/options/{optionId}/stock")
+    @RequiredRole(UserRole.STORE)
     public ResponseEntity<ProductOptionResponse> updateStock(
             @PathVariable Long storeId,
             @PathVariable Long productId,
